feat: [CODE-550]: trigger webhook harness scm (#49868)
