+++ 125-cd-nextgen/modules/common/core/src/main/java/io/harness/ng/BaseUrls.java
@@ -25,4 +25,5 @@ public class BaseUrls {
   @JsonProperty("nextGenUiUrl") String nextGenUiUrl;
   @JsonProperty("nextGenAuthUiUrl") String nextGenAuthUiUrl;
   @JsonProperty("webhookBaseUrl") String webhookBaseUrl;
+  @JsonProperty("ngManagerScmBaseUrl") String ngManagerScmBaseUrl;
 }
